ULTIMA8: Provide a message explaining why movies and maybe sound don't work
A few people have not noticed that the GOG version includes a CD image and some
files need to be taken from there.  This should help those people.

The same problem will happen for anyone who has an old install of the game and
forgets that it uses some files from the CD.